[#26] Change AuthAdapter and SurveyAdapter to just SurveyAdapter
